pavers/pavers/backend/requirements.txt
flask==2.2.5
flask-cors==3.0.10
flask-sqlalchemy==3.0.5
torch==2.0.1
opencv-python==4.7.0.72
pillow==9.5.0
numpy==1.24.3
pandas==1.5.3
matplotlib==3.7.1
requests==2.31.0